Add support AG-CX10
The fallback rcName would not work for the AG-CX10 and results in an error claiming that the "Remote setting is Disabled". Because this is definitely not the case, I analysed the traffic between the HC ROP Android app and the Camera and came across the difference in the rcName. Now, it is also possible to  connect and control the AG-CX10.